test_ops.py split into multiple files
