+++ forge/test/mlir/operators/matmul/test_matmul.py
@@ -1,0 +1,34 @@
+# SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+
+import pytest
+import torch
+from torch import nn
+
+import forge
+from forge.verify.verify import verify
+
+
+@pytest.mark.parametrize("batch_size", [1, 7, 32])
+@pytest.mark.parametrize("outer_dim_x", [7, 32, 41, 64])
+@pytest.mark.parametrize("outer_dim_y", [7, 32, 41, 64])
+@pytest.mark.parametrize("inner_dim", [1, 7, 32, 41, 64])
+@pytest.mark.push
+def test_matmul(batch_size, outer_dim_x, outer_dim_y, inner_dim):
+    class Matmul(nn.Module):
+        def __init__(self):
+            super().__init__()
+
+        def forward(self, x, y):
+            return torch.matmul(x, y)
+
+    inputs = [
+        torch.rand(batch_size, outer_dim_x, inner_dim),
+        torch.rand(batch_size, inner_dim, outer_dim_y),
+    ]
+
+    framework_model = Matmul()
+    compiled_model = forge.compile(framework_model, sample_inputs=inputs)
+
+    verify(inputs, framework_model, compiled_model)
