+++ pytest.ini
@@ -19,7 +19,7 @@ markers =
 # Where pytest should look for tests
 testpaths =
     # Ops
-    forge/test/mlir/test_ops.py
+    forge/test/mlir/operators
     forge/test/mlir/test_ops_tf.py
     forge/test/mlir/llama/tests/test_specific_ops_llama32.py
